moved the PageHeading under components. added DisplayPhotos component but commented it out for now
